add SIMD support

+++ naive.c
@@ -1,13 +1,12 @@
 #include "naive.h"
 #include <math.h>
 
-
 float naive_norm(float arr[], int n)
 {
     float res = 0.0;
     for (int i = 0; i < n; i++)
     {
-        res += sqrtf(arr[i]);
+        res += sqrt(arr[i]);
     }
     return res;
 }
